Fix bugs pertaining to read-only excel files. Also change UI from 'MailMerge' to 'Mailing Labels'.

+++ PatientRegistrationApp/Form1.cs
@@ -20,8 +20,17 @@ namespace PatientRegistrationApp
                 var app = new Excel.Application();
                 app.Visible = false;
                 m_existingWkBook = app.Workbooks.Open("Patient Registration.xlsx");
-                InitializeComponent(); // initialize app
-                mb_initialized = true;
+                if (!m_existingWkBook.ReadOnly)
+                {
+                    InitializeComponent(); // initialize app
+                    mb_initialized = true;
+                }
+                else
+                {
+                    Form prompt = new AlertDialog("The Patient Registration excel file is read-only. " +
+                        "Please configure this file correctly and reopen the application again.");
+                    prompt.ShowDialog();
+                }
             }
             catch (Exception ex)
             {
@@ -35,9 +44,37 @@ namespace PatientRegistrationApp
 
         public bool IsInitialized() { return mb_initialized; }
 
+        private bool OpenExcelFileOnAccessFailure(bool inVisible, bool inShowDialog=true)
+        {
+            bool bOutSuccess = false;
+            try
+            {
+                m_existingWkBook.Application.Visible = inVisible;
+                bOutSuccess = true;
+            }
+            catch (InvalidCastException /*ex*/)
+            {
+                var app = new Excel.Application();
+                app.Visible = inVisible;
+                m_existingWkBook = app.Workbooks.Open("Patient Registration.xlsx");
+                bOutSuccess = true;
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex.ToString());
+                if (inShowDialog)
+                {
+                    Form prompt = new AlertDialog("Failed to find Patient Registration excel file. " +
+                    "Please make sure that the excel file is in the same folder as the application.");
+                    prompt.ShowDialog();
+                }
+            }
+            return bOutSuccess;
+        }
+
         private void btnSendMailmerge_Click(object sender, EventArgs e)
         {
-            if (m_existingWkBook != null)
+            if (m_existingWkBook != null && OpenExcelFileOnAccessFailure(mb_ExcelDocumentIsOpen))
             {
                 Form mailMergeDialog = new MailMergeDialog(m_existingWkBook);
                 mailMergeDialog.ShowDialog();
@@ -46,40 +83,18 @@ namespace PatientRegistrationApp
 
         private void btnOpenXls_Click(object sender, EventArgs e)
         {
-            if (m_existingWkBook != null)
+            if (m_existingWkBook != null && OpenExcelFileOnAccessFailure(mb_ExcelDocumentIsOpen))
             {
-                try
-                {
-                    // just open excel
-                    m_existingWkBook.Application.Visible ^= true;
-                }
-                catch (InvalidCastException /*ex*/)
-                {
-                    try
-                    {
-                        var app = new Excel.Application();
-                        app.Visible = true;
-                        m_existingWkBook = app.Workbooks.Open("Patient Registration.xlsx");
-                    }
-                    catch (Exception ex)
-                    {
-                        // TODO: Show some dialog that says that it can't find excel/excel file
-                        Console.WriteLine(ex.ToString());
-                    }
-                    
-                }
-                catch (Exception ex)
-                {
-                    // TODO: Show some dialog that says that it can't find excel/excel file
-                    Console.WriteLine(ex.ToString());
-                }
+                mb_ExcelDocumentIsOpen ^= true;
+                OpenExcelFileOnAccessFailure(mb_ExcelDocumentIsOpen);
             }
         }
 
         private void btnInputPat_Click(object sender, EventArgs e)
         {
-            if (m_existingWkBook != null)
+            if (m_existingWkBook != null && OpenExcelFileOnAccessFailure(mb_ExcelDocumentIsOpen))
             {
+                OpenExcelFileOnAccessFailure(mb_ExcelDocumentIsOpen);
                 Form addPatientDialog = new AddPatientDialog(m_existingWkBook);
                 addPatientDialog.ShowDialog();
             }
@@ -87,8 +102,9 @@ namespace PatientRegistrationApp
 
         private void btnDeletePat_Click(object sender, EventArgs e)
         {
-            if (m_existingWkBook != null)
+            if (m_existingWkBook != null && OpenExcelFileOnAccessFailure(mb_ExcelDocumentIsOpen))
             {
+                OpenExcelFileOnAccessFailure(mb_ExcelDocumentIsOpen);
                 Form deletePatientDialog = new DeletePatientDialog(m_existingWkBook);
                 deletePatientDialog.ShowDialog();
             }
@@ -104,10 +120,12 @@ namespace PatientRegistrationApp
 
         protected override void OnClosed(EventArgs e)
         {
-            m_existingWkBook.Close();
+            if (OpenExcelFileOnAccessFailure(false, false))
+                m_existingWkBook.Close();
         }
 
         private bool mb_initialized = false;
+        private bool mb_ExcelDocumentIsOpen = false;
         private Excel.Workbook m_existingWkBook = null;
     }
 }
